pyproject.toml
[tool.poetry]
name = "omnibot"
version = "0.1.0"
description = ""
authors = ["Panagiotis Papaemmanouil <papaemman.pan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.7,<3.10"
streamlit = "^1.22.0"
pandas = "^2.0.1"
openpyxl = "^3.1.2"
langchain = "^0.0.167"
openai = "^0.27.6"
huggingface-hub = "^0.14.1"
setuptools = "^67.7.2"
python-dotenv = "^1.0.0"
streamlit-chat = "^0.0.2.2"
chromadb = "^0.3.22"
tiktoken = "^0.4.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
